refactor: redesign the backend crate with more logic modules
