+++ crates/backend/src/managers/idle_manager/idle_notifier.rs
