+++ crates/backend/src/managers/mod.rs
@@ -1,0 +1,2 @@
+pub(super) mod idle_manager;
+pub(super) mod window_manager;
