+++ crates/backend/src/managers/window_manager/cache.rs
